fix get_submissions (#20)
* fix get_submissions

+++ spec/test_app/db/migrate/20181114192149_create_assignment_groups.rb
@@ -2,7 +2,7 @@ class CreateAssignmentGroups < ActiveRecord::Migration[5.0]
   def change
     create_table :assignment_groups do |t|
       t.integer :course_id
-      t.integer :group_weight
+      t.float :group_weight
       t.string :name
     end
   end
